backend/requirements.txt
## Backend dependencies for Notes Taking App

Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=4.0,<5.0
